+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/toolbox/block/BlockToolboxBase.java
@@ -1,7 +1,7 @@
 package com.codetaylor.mc.artisanworktables.modules.toolbox.block;
 
+import com.codetaylor.mc.artisanworktables.ModArtisanWorktables;
 import com.codetaylor.mc.artisanworktables.modules.toolbox.tile.TileEntityToolbox;
-import com.codetaylor.mc.artisanworktables.modules.worktables.ModuleWorktables;
 import com.codetaylor.mc.artisanworktables.modules.worktables.Util;
 import com.codetaylor.mc.artisanworktables.modules.worktables.tile.spi.TileEntityBase;
 import net.minecraft.block.Block;
@@ -81,7 +81,7 @@ public abstract class BlockToolboxBase
       return true;
     }
 
-    playerIn.openGui(ModuleWorktables.MOD_INSTANCE, 1, worldIn, pos.getX(), pos.getY(), pos.getZ());
+    playerIn.openGui(ModArtisanWorktables.INSTANCE, 1, worldIn, pos.getX(), pos.getY(), pos.getZ());
     return true;
   }